[APM] refactor server folder by grouping router logic and colocate backend route files
