+++ x-pack/plugins/apm/server/index.ts
@@ -116,7 +116,7 @@ export type { APMPluginSetup } from './types';
 export type {
   APMServerRouteRepository,
   APIEndpoint,
-} from './routes/get_global_apm_server_route_repository';
+} from './routes/apm_routes/get_global_apm_server_route_repository';
 export type { APMRouteHandlerResources } from './routes/typings';
 
 export type { ProcessorEvent } from '../common/processor_event';
