+++ x-pack/plugins/apm/server/lib/helpers/create_es_client/call_async_with_debug.ts
@@ -11,7 +11,7 @@ import chalk from 'chalk';
 import { KibanaRequest } from '../../../../../../../src/core/server';
 import { RequestStatus } from '../../../../../../../src/plugins/inspector';
 import { WrappedElasticsearchClientError } from '../../../../../observability/server';
-import { inspectableEsQueriesMap } from '../../../routes/register_routes';
+import { inspectableEsQueriesMap } from '../../../routes/apm_routes/register_apm_server_routes';
 import { getInspectResponse } from '../../../../../observability/server';
 
 function formatObj(obj: Record<string, any>) {
