+++ x-pack/plugins/apm/server/routes/apm_routes/create_apm_server_route_repository.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 import { createServerRouteRepository } from '@kbn/server-route-repository';
-import { APMRouteCreateOptions, APMRouteHandlerResources } from './typings';
+import { APMRouteCreateOptions, APMRouteHandlerResources } from '../typings';
 
 export function createApmServerRouteRepository() {
   return createServerRouteRepository<
