+++ x-pack/plugins/apm/server/routes/apm_routes/get_global_apm_server_route_repository.ts
@@ -10,33 +10,33 @@ import type {
   EndpointOf,
 } from '@kbn/server-route-repository';
 import { PickByValue } from 'utility-types';
-import { alertsChartPreviewRouteRepository } from './alerts/chart_preview';
-import { backendsRouteRepository } from './backends';
-import { correlationsRouteRepository } from './correlations';
-import { createApmServerRouteRepository } from './create_apm_server_route_repository';
-import { environmentsRouteRepository } from './environments';
-import { errorsRouteRepository } from './errors';
-import { apmFleetRouteRepository } from './fleet';
-import { dataViewRouteRepository } from './data_view';
-import { latencyDistributionRouteRepository } from './latency_distribution';
-import { metricsRouteRepository } from './metrics';
-import { observabilityOverviewRouteRepository } from './observability_overview';
-import { rumRouteRepository } from './rum_client';
-import { fallbackToTransactionsRouteRepository } from './fallback_to_transactions';
-import { serviceRouteRepository } from './services';
-import { serviceMapRouteRepository } from './service_map';
-import { serviceNodeRouteRepository } from './service_nodes';
-import { agentConfigurationRouteRepository } from './settings/agent_configuration';
-import { anomalyDetectionRouteRepository } from './settings/anomaly_detection';
-import { apmIndicesRouteRepository } from './settings/apm_indices';
-import { customLinkRouteRepository } from './settings/custom_link';
-import { sourceMapsRouteRepository } from './source_maps';
-import { traceRouteRepository } from './traces';
-import { transactionRouteRepository } from './transactions';
-import { APMRouteHandlerResources } from './typings';
-import { historicalDataRouteRepository } from './historical_data';
-import { eventMetadataRouteRepository } from './event_metadata';
-import { suggestionsRouteRepository } from './suggestions';
+import { correlationsRouteRepository } from '../correlations';
+import { alertsChartPreviewRouteRepository } from '../alerts/chart_preview';
+import { backendsRouteRepository } from '../backends/route';
+import { createApmServerRouteRepository } from '../apm_routes/create_apm_server_route_repository';
+import { environmentsRouteRepository } from '../environments';
+import { errorsRouteRepository } from '../errors';
+import { apmFleetRouteRepository } from '../fleet';
+import { dataViewRouteRepository } from '../data_view';
+import { latencyDistributionRouteRepository } from '../latency_distribution';
+import { metricsRouteRepository } from '../metrics';
+import { observabilityOverviewRouteRepository } from '../observability_overview';
+import { rumRouteRepository } from '../rum_client';
+import { fallbackToTransactionsRouteRepository } from '../fallback_to_transactions';
+import { serviceRouteRepository } from '../services';
+import { serviceMapRouteRepository } from '../service_map';
+import { serviceNodeRouteRepository } from '../service_nodes';
+import { agentConfigurationRouteRepository } from '../settings/agent_configuration';
+import { anomalyDetectionRouteRepository } from '../settings/anomaly_detection';
+import { apmIndicesRouteRepository } from '../settings/apm_indices';
+import { customLinkRouteRepository } from '../settings/custom_link';
+import { sourceMapsRouteRepository } from '../source_maps';
+import { traceRouteRepository } from '../traces';
+import { transactionRouteRepository } from '../transactions';
+import { APMRouteHandlerResources } from '../typings';
+import { historicalDataRouteRepository } from '../historical_data';
+import { eventMetadataRouteRepository } from '../event_metadata';
+import { suggestionsRouteRepository } from '../suggestions';
 
 const getTypedGlobalApmServerRouteRepository = () => {
   const repository = createApmServerRouteRepository()
