+++ x-pack/plugins/apm/server/routes/apm_routes/register_apm_server_routes.test.ts
@@ -12,7 +12,7 @@ import * as t from 'io-ts';
 import { CoreSetup, Logger } from 'src/core/server';
 import { APMConfig } from '../..';
 import { APMRouteCreateOptions, APMRouteHandlerResources } from '../typings';
-import { registerRoutes } from './index';
+import { registerRoutes } from './register_apm_server_routes';
 
 type RegisterRouteDependencies = Parameters<typeof registerRoutes>[0];
