+++ x-pack/plugins/apm/server/routes/backends/get_metadata_for_backend.ts
@@ -9,7 +9,7 @@ import { maybe } from '../../../common/utils/maybe';
 import { ProcessorEvent } from '../../../common/processor_event';
 import { SPAN_DESTINATION_SERVICE_RESOURCE } from '../../../common/elasticsearch_fieldnames';
 import { rangeQuery } from '../../../../observability/server';
-import { Setup } from '../helpers/setup_request';
+import { Setup } from '../../lib/helpers/setup_request';
 
 export async function getMetadataForBackend({
   setup,
