+++ x-pack/plugins/apm/server/routes/backends/get_top_backends.ts
@@ -8,9 +8,9 @@
 import { kqlQuery } from '../../../../observability/server';
 import { NodeType } from '../../../common/connections';
 import { environmentQuery } from '../../../common/utils/environment_query';
-import { getConnectionStats } from '../connections/get_connection_stats';
-import { getConnectionStatsItemsWithRelativeImpact } from '../connections/get_connection_stats/get_connection_stats_items_with_relative_impact';
-import { Setup } from '../helpers/setup_request';
+import { getConnectionStats } from '../../lib/connections/get_connection_stats';
+import { getConnectionStatsItemsWithRelativeImpact } from '../../lib/connections/get_connection_stats/get_connection_stats_items_with_relative_impact';
+import { Setup } from '../../lib/helpers/setup_request';
 
 export async function getTopBackends({
   setup,
