+++ x-pack/plugins/apm/server/routes/backends/route.ts
@@ -7,16 +7,21 @@
 
 import * as t from 'io-ts';
 import { toNumberRt } from '@kbn/io-ts-utils/to_number_rt';
-import { setupRequest } from '../lib/helpers/setup_request';
-import { environmentRt, kueryRt, offsetRt, rangeRt } from './default_api_types';
-import { createApmServerRoute } from './create_apm_server_route';
-import { createApmServerRouteRepository } from './create_apm_server_route_repository';
-import { getMetadataForBackend } from '../lib/backends/get_metadata_for_backend';
-import { getLatencyChartsForBackend } from '../lib/backends/get_latency_charts_for_backend';
-import { getTopBackends } from '../lib/backends/get_top_backends';
-import { getUpstreamServicesForBackend } from '../lib/backends/get_upstream_services_for_backend';
-import { getThroughputChartsForBackend } from '../lib/backends/get_throughput_charts_for_backend';
-import { getErrorRateChartsForBackend } from '../lib/backends/get_error_rate_charts_for_backend';
+import { setupRequest } from '../../lib/helpers/setup_request';
+import {
+  environmentRt,
+  kueryRt,
+  offsetRt,
+  rangeRt,
+} from '../default_api_types';
+import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
+import { createApmServerRouteRepository } from '../apm_routes/create_apm_server_route_repository';
+import { getMetadataForBackend } from './get_metadata_for_backend';
+import { getLatencyChartsForBackend } from './get_latency_charts_for_backend';
+import { getTopBackends } from './get_top_backends';
+import { getUpstreamServicesForBackend } from './get_upstream_services_for_backend';
+import { getThroughputChartsForBackend } from './get_throughput_charts_for_backend';
+import { getErrorRateChartsForBackend } from './get_error_rate_charts_for_backend';
 
 const topBackendsRoute = createApmServerRoute({
   endpoint: 'GET /internal/apm/backends/top_backends',
