+++ x-pack/plugins/apm/server/routes/event_metadata.ts
@@ -6,8 +6,8 @@
  */
 
 import * as t from 'io-ts';
-import { createApmServerRouteRepository } from './create_apm_server_route_repository';
-import { createApmServerRoute } from './create_apm_server_route';
+import { createApmServerRouteRepository } from './apm_routes/create_apm_server_route_repository';
+import { createApmServerRoute } from './apm_routes/create_apm_server_route';
 import { getEventMetadata } from '../lib/event_metadata/get_event_metadata';
 import { processorEventRt } from '../../common/processor_event';
 import { setupRequest } from '../lib/helpers/setup_request';
