+++ x-pack/plugins/apm/server/routes/fallback_to_transactions.ts
@@ -8,8 +8,8 @@
 import * as t from 'io-ts';
 import { getIsUsingTransactionEvents } from '../lib/helpers/transactions/get_is_using_transaction_events';
 import { setupRequest } from '../lib/helpers/setup_request';
-import { createApmServerRoute } from './create_apm_server_route';
-import { createApmServerRouteRepository } from './create_apm_server_route_repository';
+import { createApmServerRoute } from './apm_routes/create_apm_server_route';
+import { createApmServerRouteRepository } from './apm_routes/create_apm_server_route_repository';
 import { kueryRt, rangeRt } from './default_api_types';
 
 const fallbackToTransactionsRoute = createApmServerRoute({
