+++ x-pack/plugins/apm/server/routes/historical_data/index.ts
@@ -6,8 +6,8 @@
  */
 
 import { setupRequest } from '../../lib/helpers/setup_request';
-import { createApmServerRoute } from '../create_apm_server_route';
-import { createApmServerRouteRepository } from '../create_apm_server_route_repository';
+import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
+import { createApmServerRouteRepository } from '../apm_routes/create_apm_server_route_repository';
 import { hasHistoricalAgentData } from './has_historical_agent_data';
 
 const hasDataRoute = createApmServerRoute({
