+++ x-pack/plugins/apm/server/routes/latency_distribution.ts
@@ -9,8 +9,8 @@ import * as t from 'io-ts';
 import { toNumberRt } from '@kbn/io-ts-utils/to_number_rt';
 import { getOverallLatencyDistribution } from '../lib/latency/get_overall_latency_distribution';
 import { setupRequest } from '../lib/helpers/setup_request';
-import { createApmServerRoute } from './create_apm_server_route';
-import { createApmServerRouteRepository } from './create_apm_server_route_repository';
+import { createApmServerRoute } from './apm_routes/create_apm_server_route';
+import { createApmServerRouteRepository } from './apm_routes/create_apm_server_route_repository';
 import { environmentRt, kueryRt, rangeRt } from './default_api_types';
 
 const latencyOverallDistributionRoute = createApmServerRoute({
