+++ x-pack/plugins/apm/server/routes/settings/anomaly_detection.ts
@@ -10,7 +10,7 @@ import Boom from '@hapi/boom';
 import { maxSuggestions } from '../../../../observability/common';
 import { isActivePlatinumLicense } from '../../../common/license_check';
 import { ML_ERRORS } from '../../../common/anomaly_detection';
-import { createApmServerRoute } from '../create_apm_server_route';
+import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
 import { getAnomalyDetectionJobs } from '../../lib/anomaly_detection/get_anomaly_detection_jobs';
 import { createAnomalyDetectionJobs } from '../../lib/anomaly_detection/create_anomaly_detection_jobs';
 import { setupRequest } from '../../lib/helpers/setup_request';
@@ -19,7 +19,7 @@ import { hasLegacyJobs } from '../../lib/anomaly_detection/has_legacy_jobs';
 import { getSearchAggregatedTransactions } from '../../lib/helpers/transactions';
 import { notifyFeatureUsage } from '../../feature';
 import { withApmSpan } from '../../utils/with_apm_span';
-import { createApmServerRouteRepository } from '../create_apm_server_route_repository';
+import { createApmServerRouteRepository } from '../apm_routes/create_apm_server_route_repository';
 
 // get ML anomaly detection jobs for each environment
 const anomalyDetectionJobsRoute = createApmServerRoute({
